+++ src/modules/site/src/ProjectAccessControlHandler.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace Drupal\site;
+
+use Drupal\Core\Access\AccessResult;
+use Drupal\Core\Entity\EntityAccessControlHandler;
+use Drupal\Core\Entity\EntityInterface;
+use Drupal\Core\Session\AccountInterface;
+
+/**
+ * Defines the access control handler for the project entity type.
+ */
+class ProjectAccessControlHandler extends EntityAccessControlHandler {
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function checkAccess(EntityInterface $entity, $operation, AccountInterface $account) {
+
+    switch ($operation) {
+      case 'view':
+        return AccessResult::allowedIfHasPermission($account, 'view project');
+
+      case 'update':
+        return AccessResult::allowedIfHasPermissions(
+          $account,
+          ['edit project', 'administer project'],
+          'OR',
+        );
+
+      case 'delete':
+        return AccessResult::allowedIfHasPermissions(
+          $account,
+          ['delete project', 'administer project'],
+          'OR',
+        );
+
+      default:
+        // No opinion.
+        return AccessResult::neutral();
+    }
+
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function checkCreateAccess(AccountInterface $account, array $context, $entity_bundle = NULL) {
+    return AccessResult::allowedIfHasPermissions(
+      $account,
+      ['create project', 'administer project'],
+      'OR',
+    );
+  }
+
+}