+++ src/modules/site/src/ProjectTypeListBuilder.php
@@ -1,0 +1,51 @@
+<?php
+
+namespace Drupal\site;
+
+use Drupal\Core\Config\Entity\ConfigEntityListBuilder;
+use Drupal\Core\Entity\EntityInterface;
+use Drupal\Core\Url;
+
+/**
+ * Defines a class to build a listing of project type entities.
+ *
+ * @see \Drupal\site\Entity\ProjectType
+ */
+class ProjectTypeListBuilder extends ConfigEntityListBuilder {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function buildHeader() {
+    $header['title'] = $this->t('Label');
+
+    return $header + parent::buildHeader();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function buildRow(EntityInterface $entity) {
+    $row['title'] = [
+      'data' => $entity->label(),
+      'class' => ['menu-label'],
+    ];
+
+    return $row + parent::buildRow($entity);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function render() {
+    $build = parent::render();
+
+    $build['table']['#empty'] = $this->t(
+      'No project types available. <a href=":link">Add project type</a>.',
+      [':link' => Url::fromRoute('entity.project_type.add_form')->toString()]
+    );
+
+    return $build;
+  }
+
+}