+++ src/modules/site/src/SiteBreadcrumbBuilder.php
@@ -22,7 +22,7 @@ class SiteBreadcrumbBuilder implements BreadcrumbBuilderInterface {
    * {@inheritdoc}
    */
   public function applies(RouteMatchInterface $route_match) {
-    $project = $route_match->getParameter('drupal_project');
+    $project = $route_match->getParameter('project');
     $site = $route_match->getParameter('site');
     return $project instanceof DrupalProjectInterface || $site instanceof SiteEntity;
   }
@@ -36,17 +36,17 @@ class SiteBreadcrumbBuilder implements BreadcrumbBuilderInterface {
     $breadcrumb->addLink(Link::createFromRoute($this->t('Home'), '<front>'));
 
     $site = $route_match->getParameter('site');
-    $project = $route_match->getParameter('drupal_project');
+    $project = $route_match->getParameter('project');
 
     // If "Projects" view exists, set as parent breadcrumb.
     try {
-      $projects_route = 'view.drupal_projects.page_1';
+      $projects_route = 'view.projects.page_1';
       \Drupal::service('router.route_provider')
         ->getRouteByName($projects_route);
       ;
     }
     catch (\Exception $e) {
-      $projects_route = 'entity.drupal_project.collection';
+      $projects_route = 'entity.project.collection';
     }
 
     // If "Sites" view exists, set as parent breadcrumb.
@@ -64,7 +64,7 @@ class SiteBreadcrumbBuilder implements BreadcrumbBuilderInterface {
     if ($project) {
       $breadcrumb->addLink(Link::createFromRoute($this->t('Projects'), $projects_route));
 
-      if (\Drupal::routeMatch()->getRouteName() == 'entity.drupal_project.add_site') {
+      if (\Drupal::routeMatch()->getRouteName() == 'entity.project.add_site') {
         $breadcrumb->addLink($project->toLink());
       }
 
@@ -72,10 +72,10 @@ class SiteBreadcrumbBuilder implements BreadcrumbBuilderInterface {
     }
 
     // If on a site with a drupal project, use Projects > Project Name
-    if (!empty($site->drupal_project->entity)) {
+    if (!empty($site->project->entity)) {
       $breadcrumb->addLink(Link::createFromRoute($this->t('Projects'), $projects_route));
-      $breadcrumb->addLink($site->drupal_project->entity->toLink());
-      $breadcrumb->addLink($site->drupal_project->entity->toLink(t('Environments')));
+      $breadcrumb->addLink($site->project->entity->toLink());
+      $breadcrumb->addLink($site->project->entity->toLink(t('Environments')));
     }
     // All other sites use Sites
     else {